--- PETsARD/synthesizer/sdv/__init__.py
@@ -1,6 +1,0 @@
-from PETsARD.synthesizer.sdv.sdv import SDV
-
-
-__all__ = [
-    'SDV'
-]
